# packages/http-api/src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Feedless HTTP API
  description: REST API for Feedless entities (auth, repository, source, document, harvest, plan, group).
  version: 1.0.0
servers:
  - url: '{scheme}://{host}/api/v1'
    description: Feedless HTTP API v1. Feedless is self-hosted — point `host` at your instance.
    variables:
      scheme:
        default: https
        enum: [https, http]
      host:
        default: feedless.org
        description: Host (and optional port) of the Feedless instance
security:
  - bearerAuth: []
paths:
...
  /records:
    get:
      tags: [Records]
      operationId: listRecords
      parameters:
        - name: repositoryId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordListResponse'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Records]
      operationId: createRecord
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Record'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags: [Records]
      operationId: deleteRecords
      parameters:
        - name: repositoryId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordDeleteRequest'
      responses:
        '204':
          description: Deleted
        '429':
          $ref: '#/components/responses/TooManyRequests'
        default:
          $ref: '#/components/responses/Error'
  /records/{recordId}:
    parameters:
      - $ref: '#/components/parameters/RecordId'
    get:
      tags: [Records]
      operationId: getRecord
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Record'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
    patch:
      tags: [Records]
      operationId: updateRecord
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Record'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        default:
          $ref: '#/components/responses/Error'
  /plans:
    get:
      tags: [Plans]
      operationId: listPlans
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanListResponse'
        default:
          $ref: '#/components/responses/Error'
  /plans/{planId}:
    get:
      tags: [Plans]
      operationId: getPlan
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plan'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
  /groups:
    get:
      tags: [Groups]
      operationId: listGroups
      responses:
        '200':
          description: Groups the current user belongs to (as GroupAssignment)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupAssignmentListResponse'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Groups]
      operationId: createGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        default:
          $ref: '#/components/responses/Error'
  /groups/{groupId}:
    parameters:
      - name: groupId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Groups]
      operationId: getGroup
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags: [Groups]
      operationId: deleteGroup
      responses:
        '204':
          description: Deleted
        '429':
          $ref: '#/components/responses/TooManyRequests'
        default:
          $ref: '#/components/responses/Error'
  /groups/{groupId}/members:
    parameters:
      - name: groupId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Groups]
      operationId: listGroupMembers
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Members of the group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupMemberListResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Groups]
      operationId: addGroupMember
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupMemberCreate'
      responses:
        '201':
          description: Assignment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupAssignment'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        default:
          $ref: '#/components/responses/Error'
  /groups/{groupId}/members/{userId}:
    parameters:
      - name: groupId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      tags: [Groups]
      operationId: removeGroupMember
      responses:
        '204':
          description: Removed
        '429':
          $ref: '#/components/responses/TooManyRequests'
        default:
          $ref: '#/components/responses/Error'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        `Authorization: Bearer <UserSecret.value>` (GitHub PAT style).
        Legacy `Authentication: Bearer <…>` accepted in v1 but **deprecated**.
        TOKEN cookie still accepted for browser clients. Anonymous JWTs are rejected.
  parameters:
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 0
        maximum: 100
        default: 20
...
    RecordId:
      name: recordId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Error:
      description: >
        Any error. Every error response — 400 `BAD_REQUEST` / `VALIDATION_ERROR`,
        401 `UNAUTHORIZED`, 403 `FORBIDDEN`, 404 `NOT_FOUND`, 429 `TOO_MANY_REQUESTS`,
        500 `INTERNAL_ERROR` — carries an `ApiError` body. Since every route requires
        authentication, every operation can return 401.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    TooManyRequests:
      description: Rate limit exceeded — retry after the interval in `Retry-After`
      headers:
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine-readable code, e.g. NOT_FOUND, VALIDATION_ERROR
        message:
          type: string
          description: Human-readable summary. Not stable — do not match on it.
        corrId:
          type: string
          description: Correlation id for support/debug
        path:
          type: string
          description: Request path the error occurred on
        errors:
          type: array
          description: Per-field details, present on VALIDATION_ERROR. All failing fields, not just the first.
          items:
            $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
...
    Record:
      type: object
      required: [id, url, createdAt, publishedAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
        title:
          type: string
        text:
          type: string
        html:
          type: string
        tags:
          type: array
          items:
            type: string
        imageUrl:
          type: string
        createdAt:
          type: string
          format: date-time
        publishedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        startingAt:
          type: string
          format: date-time
        rawBase64:
          type: string
        rawMimeType:
          type: string
    RecordCreate:
      type: object
      required: [title, url, publishedAt, repositoryId]
      properties:
        title:
          type: string
        url:
          type: string
        publishedAt:
          type: string
          format: date-time
        repositoryId:
          type: string
          format: uuid
        text:
          type: string
        tags:
          type: array
          items:
            type: string
        rawBase64:
          type: string
        rawMimeType:
          type: string
    RecordUpdate:
      type: object
      properties:
        title:
          type: string
        text:
          type: string
        url:
          type: string
        tags:
          type: array
          items:
            type: string
    RecordDeleteRequest:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          items:
            type: string
            format: uuid
    RecordListResponse:
      type: object
      required: [items, hasMore]
      properties:
        hasMore:
          type: boolean
        items:
          type: array
          items:
            $ref: '#/components/schemas/Record'
...
    Plan:
      type: object
      required: [id, productId, recurringYearly]
      properties:
        id:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        startedAt:
          type: string
          format: date-time
        terminatedAt:
          type: string
          format: date-time
        recurringYearly:
          type: boolean
    PlanListResponse:
      type: object
      required: [items, hasMore]
      properties:
        hasMore:
          type: boolean
        items:
          type: array
          items:
            $ref: '#/components/schemas/Plan'
    Group:
      type: object
      required: [id, name, ownerId]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        ownerId:
          type: string
          format: uuid
    GroupCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
    GroupAssignment:
      type: object
      required: [id, role, name]
      properties:
        id:
          type: string
          format: uuid
        role:
          type: string
          enum: [owner, viewer, editor]
        name:
          type: string
    GroupAssignmentListResponse:
      type: object
      required: [items, hasMore]
      properties:
        hasMore:
          type: boolean
        items:
          type: array
          items:
            $ref: '#/components/schemas/GroupAssignment'
    GroupMember:
      type: object
      required: [userId, role]
      properties:
        userId:
          type: string
          format: uuid
        role:
          type: string
          enum: [owner, viewer, editor]
    GroupMemberListResponse:
      type: object
      required: [items, hasMore]
      properties:
        hasMore:
          type: boolean
        items:
          type: array
          items:
            $ref: '#/components/schemas/GroupMember'
    GroupMemberCreate:
      type: object
      required: [userId, role]
      properties:
        userId:
          type: string
          format: uuid
        role:
          type: string
          enum: [owner, viewer, editor]